requirements.txt
langchain
PyPDF2
python-dotenv
openai
faiss-cpu
tiktoken
beautifulsoup4
fastapi
uvicorn
fastapi-sessions
python-multipart
